Addition of Debug Logs
Added debug logs. The format of adding the debug log should be as follow:
this.getName() + " - " + Msg
A further improvement that we can make is that we can also add the function names in the debug logs. For that we can use different methods for example we can use
Thread.currentThread().getStackTrace()[1].getMethodName()
this returns the function name and we can place it after getName() and before message so that we can have a better traceability of the debug log messages.